Add SelectionModeManager
The SelectionModeManager is activated by Shift-S (by default) and
essentially provides a utility class to manage a number of different
selection behaviors that are toggled by user-configurable hotkeys.

This is the first major part of the implementation of Feature #4268,
Multi-Geometry Modal Selection and Deselection, as proposed by
FreeCAD forums member bgoodr.