Modify time-handling to not use shared statics
The C time functions localtime, asctime, and ctime use a shared
static time structure that is overwritten whenever one of them
is called. LGTM points out this flaw and suggests replacing them
in modern code with a safer alternative. This change does that,
using versions of those functions that take a local time structure
as an argument.

To avoid introducing the same flaw in our own code, the Time()
function of ConsoleSingleton is modified to return a new std::string
object created from the resulting time string, rather than a pointer
to memory that may be overwritten by later calls.