[0.19] HiDPI/Ultrawide font size correction
When using QtWebEngine it is not necessary or desirable to manually
upscale the minimum font size in a web display. The hard-coded 1920
pixel "standard" size results in undesirable behavior on HiDPI and
ultrawide screens when using the newer QtWebEngine, which correctly
interprets "px" as 1/96 of an inch (rather than device pixels). If any
scaling is required it should be applied via a CSS media query in the
web content.

Note that this patch (and the bug it fixes) only affects users with a
set minimum font size. In the default case the minimum size is zero, so
is unaffected by the applied scale factor in any case.